+++ src/pages/add_expense_page.cpp
@@ -108,7 +108,7 @@ add_expense_page::validate_date(const std::string &inp) {
 
   second.day = static_cast<uint8_t>(std::stoi(match[1].str()));
   second.month = static_cast<uint8_t>(std::stoi(match[2].str()));
-  second.year = static_cast<uint8_t>(std::stoi(match[3].str()));
+  second.year = static_cast<uint16_t>(std::stoi(match[3].str()));
 
   if (!second.is_valid()) {
     return std::make_pair("Invalid date range.\n", second);
